+++ test/specs/options/autoplay.spec.js
@@ -1,0 +1,8 @@
+describe('autoplay (option)', () => {
+  it('should be true by default', () => {
+    const image = window.createImage();
+    const viewer = new Viewer(image);
+
+    expect(viewer.options.autoplay).to.be.true;
+  });
+});